Delete threads that do not meet activity requirements (#137)
Ticket: [MWAY-15](https://naurffxiv.atlassian.net/browse/MWAY-15)
Adds a task that runs once every 24 hours that checks threads for
activity requirements. Additionally, the task also checks if the starter
message is present. If the thread doesn't meet either requirements, it's
deleted and a DM will be sent (commented out for now to deal with
backlog present in NAUR) to the original owner. A log will also be made
upon each deletion.

In debug mode, it will check for the `INACTIVE_FORUM_CHANNEL_ID` and
`INACTIVE_FORUM_DURATION` environment variables.

# Log message

![image](https://github.com/user-attachments/assets/6083dcee-6eb9-4586-b07b-c2b325ba9f2a)

# DM

![image](https://github.com/user-attachments/assets/5946ce98-03ea-4f0c-95db-7ecba1a3d45b)


[MWAY-15]:
https://naurffxiv.atlassian.net/browse/MWAY-15?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ