Update timestamp for cooldown period (#161)
Ticket: MOD-76

Current timestamp is retrieved along with the amount of seconds left on
cooldown. The output adds both values together to get the unix timestamp
for the end of the cooldown period. The same line of code formats the
message into the markdown format discord allows showing how much time is
remaining regardless of days, hours, or seconds.

Reference for unix timestamp discord capability:
https://gist.github.com/LeviSnoot/d9147767abeef2f770e9ddcd91eb85aa