Take out excessive user in db check for unexile
Ticket: MOD-81
Bot returns an error if target trying to get unexiled does not exist in database.
Note: I am not sure if there is a condition where user exists in users table but not in the exiles so I could not get the next if statement to trigger. Please review.